modules/ai-auth/requirements.txt
# Authentication & Authorization Module Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
cryptography==41.0.8
bcrypt==4.1.2
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
